fpga: make reads via SPI always 32-bit aligned and 32-bit sized

+++ esp32/main/scope.h
@@ -19,7 +19,7 @@ void scope__set_triggers(uint32_t mask,
 
 void scope__get_capture_info(uint32_t *status, uint32_t *counter, uint32_t *trig_address);
 // Get 64 samples (256 bytes).
-void scope__get_capture_data_block64(uint8_t channel, uint8_t offset, uint8_t*dest);
+void scope__get_capture_data_block64(uint8_t channel, uint8_t offset, uint32_t*dest);
 
 int scope__get_group_size(scope_group_t group);
 const char *scope__get_signal_name(scope_group_t group, int index);
